This commit renames the "Emergency Services" feature to "Services" in the tenant portal to better reflect its general-purpose nature.
The following changes were made:
- Renamed `EmergencyServices.tsx` page and component to `Services.tsx`.
- Updated the route in `App.tsx` from `/tenant/emergency-services` to `/tenant/services`.
- Updated the navigation link in the tenant sidebar from "Emergency Services" to "Services".
- Updated the title and description on the Services page.

The request to add a "Call Now" button to each service was verified, and this functionality was already present in the existing component.
